LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-remove-expected.txt


FAIL First meta applies. assert_equals: expected "dark" but got "light"
PASS Second meta applies after first one is removed.
PASS Initial color-scheme with both meta elements removed.
